feat: Stop bad imports, don't update backup.go
